Add VM architecture, composition, and test infrastructure
- Add architecture module with exec/driver/kernel separation
- Add composition module for nested VM support
- Add determinism and internalization modules
- Add exec API and output condition handling
- Expand lean-bridge with invariant verification and VM export
- Add schema versioning and property-based tests
- Add threaded lane runtime and composition stress tests
- Add cross-target matrix and differential step tests
- Update existing tests for new VM structure

+++ rust/simulator/src/material.rs
@@ -56,7 +56,7 @@ pub struct HamiltonianParams {
     /// Particle mass.
     pub mass: f64,
     /// Number of spatial dimensions per particle (1 for this 2-body case).
-    pub dimensions: usize,
+    pub dimensions: u32,
     /// Initial positions for each role (length = number of roles).
     pub initial_positions: Vec<f64>,
     /// Initial momenta for each role (length = number of roles).
@@ -78,7 +78,7 @@ pub struct ContinuumFieldParams {
     /// Kernel coupling strength K. Controls diffusion rate between sites.
     pub coupling: f64,
     /// Number of field components per site (1 for scalar field).
-    pub components: usize,
+    pub components: u32,
     /// Initial field values per role (length = number of roles).
     pub initial_fields: Vec<f64>,
     /// Integration step size.
